requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
Flask>=2.0.1
numpy>=1.21.0
scikit-learn>=0.24.2 